Move all the things to separate classes
Because of ugly visibility reasons, we need to pass `this` into
PagingIterable.  Will try to extract some sensible class next.
